tagservice and controller ready

+++ giftcertificates/api/src/main/resources/sql/3_create_tables.sql
@@ -2,7 +2,7 @@ use giftcertificates;
 
 CREATE TABLE tag
 (
-    id integer not null auto_increment,
+    id bigint not null auto_increment,
     name varchar(100),
     CONSTRAINT pk_tag PRIMARY KEY (id),
     CONSTRAINT uk_tag UNIQUE (name)
@@ -10,7 +10,7 @@ CREATE TABLE tag
 
 CREATE TABLE gift_certificate
 (
-    id             integer not null auto_increment,
+    id             bigint not null auto_increment,
     name           varchar(255),
     description    varchar(500),
     price          decimal,
@@ -22,8 +22,8 @@ CREATE TABLE gift_certificate
 
 CREATE TABLE gift_certificate_tag
 (
-    tag_id              integer,
-    gift_certificate_id integer,
+    tag_id              bigint,
+    gift_certificate_id bigint,
     CONSTRAINT fk_tag_id FOREIGN KEY (tag_id) REFERENCES tag (id),
     CONSTRAINT fk_gift_certificate_id FOREIGN KEY (gift_certificate_id) REFERENCES gift_certificate (id)
 );
